Fixes issue 1612. Put outgoing call chooser in separate task affinity and exclude it from recent.
git-svn-id: http://csipsimple.googlecode.com/svn/trunk/CSipSimple@1301 9f815046-5998-e9c0-b7e2-ac03a23edfa4

+++ src/com/csipsimple/db/AccountAdapter.java
@@ -139,7 +139,7 @@ public class AccountAdapter extends ArrayAdapter<SipProfile> implements OnClickL
 	public void onClick(View v) {
 		if(service != null) {
 			try {
-				service.setAccountRegistration((Integer)v.getTag(), 1);
+				service.setAccountRegistration(((Long)v.getTag()).intValue(), 1);
 			} catch (RemoteException e) {
 				Log.e(THIS_FILE, "Unable to contact service", e);
 			}
